feat(io.reader): draft implementation of ITK/ANTs' composite HDF5 transforms
Uses ``h5py`` to browse the transforms group, and concatenates them in a list
ready for nitransforms' wrapping into a ``TransformsChain``.

The writer is not included in this PR.

The implementation takes into account the misspelling of "Tranforms" present
in transforms generated with older versions of ITK.

This implementation only accepts the composite transforms generated by ANTs.

Resolves: #4