chore: Use setuptools_scm to manage versioning
Summary
-------
Pilots whether we can abandon versioneer in all projects, as
setuptools_scm has minimal boilerplate.

Changes
-------
The setuptools_scm module is added to the pyproject.toml file,
as recommended in the documentation.
A tiny hack is added to the root ``__init__.py`` file, so that
the version is read from ``nitransforms/_version.py`` when that
file exists or it is interpolated via ``setuptools_scm.get_version``.

To make the version traverse into container images, before building,
``python setup.py --version`` must be invoked first (and it will create
the ``nitransforms/_version.py`` file.

``nitransforms/_version.py`` has been added to ``.gitignore`` to avoid
adding it by mistake.

Finally, the CircleCI workflow is also modified, so that the packaging
is always tested and installed versions are checked both for python
packages and docker distributions.

Resolves: #82